requirements.txt
feedparser==6.0.11
pyyaml==6.0.2
jinja2==3.1.4
nltk==3.8.1
